==> scripts/30_blobcity_index_scraper.ipynb <==
# imports
import requests
import json
import os
import platform
import time
import datetime
import pandas as pd
import re

from bs4 import BeautifulSoup
from selenium import webdriver
from selenium.webdriver.support.ui import WebDriverWait
from selenium.webdriver.support import expected_conditions as EC
from selenium.webdriver.common.by import By
from selenium.common.exceptions import TimeoutException
from selenium.webdriver.chrome.options import Options

# generic store data to file function
def store_data(data, file, mode='w', toJson=False):
    if toJson:
        data = json.dumps(data)
    with open(file, mode, encoding='utf-8') as fp:
        result = fp.write(data)
        return result
    
# generic load data from file function
def load_data(file, fromJson=False):
    if os.path.isfile(file):
        with open(file, 'r', encoding='utf-8', errors="ignore") as fp:
            data = fp.read()
            if fromJson:
                data = json.loads(data)
            return data
    else:
        return 'file not found'

# test text
#print(store_data('Hello', '../data/repositories/mlart/test.txt'))
#print(load_data('../data/repositories/mlart/test.txt'))

# test json
#print(store_data({'msg':'Hello World'}, '../data/repositories/mlart/test.json', toJson=True))
#print(load_data('../data/repositories/mlart/test.json', fromJson=True))

# blobcity index scraper
def url_scraper(start = 0,
                end = 0,
                windowSize = "1280,720",
                headless = True,
                quitOnEnd = True,
                waitBetweenPress = 30,
                waitUntilTimeout = 20,
                wait = 1,
               ):
    
    url = 'https://cloud.blobcity.com/#/ps/explore'
    store = '../data/repositories/blobcity/index/'
    
    pages = {}
    
    if start == 0 and end == 0:
        return None
    
    if start > end:
        return None
    
    
    # selenium version
    CHROMEDRIVER_PATH = './.selenium/chromedriver'
    WINDOW_SIZE = windowSize

    chrome_options = Options()
    if headless:
        chrome_options.add_argument("--headless")
    chrome_options.add_argument("--window-size=%s" % WINDOW_SIZE)

    driver = webdriver.Chrome(executable_path=CHROMEDRIVER_PATH, options=chrome_options)
    driver.implicitly_wait(waitBetweenPress)
    driver.get(url)

    # wait for a given class to be loaded (javascript generated code)
    try:
        # not-found
        element = WebDriverWait(driver, waitUntilTimeout).until(EC.visibility_of_element_located((By.CLASS_NAME, 'show-for-sr')))
        print ("Page is ready!")
    except TimeoutException:
        print ("class - loading took too much time!")
        

            
    running = True
    i = start
        
    # store first page too
    if start == 1:
        time.sleep(wait)
        html = driver.page_source
        print(i, len(html))
        store_data(html, store+str(i)+'.html')
        i += 1

    
    while running:
        
        pressLink = str(i)

        # press link
        if pressLink != '':
            try:
                print('link', pressLink)
                element = WebDriverWait(driver, waitUntilTimeout).until(EC.visibility_of_element_located((By.LINK_TEXT, pressLink)))
                element.click()
            except TimeoutException:
                print ("Loading took too much time!")

        html = driver.page_source
        print(i, len(html))
        time.sleep(wait)
        store_data(html, store+str(i)+'.html')
        
        i += 1
        if i > end:
            running = False

            
    if quitOnEnd:
        time.sleep(5)
        driver.close()
        print('done')



url_scraper(start = 1, end = 125, headless = False, quitOnEnd = True)

# scrape index

store = '../data/repositories/blobcity/index/'
fp_csv = '../data/database/blobcity_01_index.csv'


files = os.listdir(store)

def scrape_index(html):
    soup = BeautifulSoup(html, 'html.parser')
    result = []
    
    base_url = 'https://cloud.blobcity.com/'
    
    items = soup.select('nb-card')
    
    for item in items:
        meta = {}
        
        header = item.find('nb-card-header')
        meta['title'] = header.text.strip()
        meta['blobcity_link'] = base_url + header.find('a').get('href')
        
        body = item.find('nb-card-body')
        meta['description'] = body.find('div', class_='col-md-7').text.strip()
        meta['github_link'] = str(body.find('a'))
        meta['github_link'] = re.findall(r'<a[^>]* href="([^"]*)"', meta['github_link'])
        if len(meta['github_link']) > 0:
            meta['github_link'] = meta['github_link'][0]
        else:
            meta['github_link'] = ''
        
        footer = item.find('nb-card-footer')
        footer = footer.select('div.col-md-6')
        meta['tags'] = footer[0].text.strip()
        meta['tags'] = meta['tags'].split(' • ')
        dates = footer[1].text.strip()
        dates = dates.split(' \xa0\xa0  ')
        # 'dates': ['Created at: 11/25/20, 11:29 AM', 'Modified at : 11/25/20, 11:29 AM']
        if (len(dates) > 0):
            meta['created_at'] = dates[0].replace('Created at: ','').strip()
            meta['modified_at'] = dates[1].replace('Modified at : ','').strip()
            
            try:
                meta['created_at'] = str(datetime.datetime.strptime(meta['created_at'], '%m/%d/%y, %I:%M %p'))
                meta['modified_at'] = str(datetime.datetime.strptime(meta['modified_at'], '%m/%d/%y, %I:%M %p'))
            except:
                print('cannot convert to datetime', meta['created_at'], meta['modified_at'])
        else:
            meta['created_at'] = footer[1].text.strip()
            meta['modified_at'] = ''
        
        result.append(meta)

    return result

items = []
for file in files:
    print(file)
    html = load_data(os.path.join(store,file))
    if 'file not found' in html:
        print(html)
        break
    meta = scrape_index(html)
    #print(meta)
    items.extend(meta)
    #break
    
print(len(items), items[0])
#store_data(links, folder_base+folder+out, toJson=True)

df = pd.DataFrame(items)
print(df.shape)

# drop duplicates
df = df.drop_duplicates(['github_link'])
print(df.shape)

# drop nan
df.dropna(subset = ['github_link'])

print(df.shape)
print(df.head())
        
# drop columns
#df.drop(columns=['author', 'submission'], inplace=True)

df.to_csv(fp_csv, sep=';', index=False)
